Add "exectty" subcommand for use when svn launches an editor.
Sometimes svn wants to launch an editor, such as when a merge conflict
occurs.  When svnwrap is capturing stdout (for example, to highlight the
output), it has stdout connected to a pipe, which is unsuitable for an
editor.  The "exectty" subcommand allows svnwrap itself to be a
front-end for the editor.  It duplicates the stderr file descriptor
(which presumably is still connected to the terminal) onto stdout, then
re-execs the passed-in command-line options.  With the present change,
svnwrap now detects the configured editor and inserts "svnwrap exectty "
in front of the original editor definition, allowing stdout to be
re-connected to the terminal.