Restore the default signal handlers for child processes.
Python sets up default handlers for several signals--which includes
ignoring SIGPIPE.  Unfortunately, these are inherited by the child
process.  When using 'diff-highlight | less' as the pager, SIGPIPE will
be ignored by diff-highlight and will try to keep reading from stdin.
If you try to exit less with the 'q' command, but have a big Subversion
operation running, then we continue to pipe data from Subversion to
diff-highlight forcing you to ctrl-c to terminate the operation.

Instead, let's restore the default signal handlers in the child
processes.  Not only does it provide a cleaner environment for the
children, it helps fix the above situation.